+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ##
 ## Added
 - creating new documents
+- user image in editor
 
 ## 1.1.0
 ## Added